pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "legal-word-drift"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "aiohttp",
    "beautifulsoup4",
    "gensim",
    "lxml",
    "matplotlib",
    "networkx",
    "numpy",
    "pandas",
    "pyarrow",
    "python-louvain",
    "rdflib",
    "scikit-learn",
    "scipy",
    "SPARQLWrapper",
    "tol-colors",
    "torch>=2.0",
    "tqdm",
    "transformers>=4.30",
    "umap-learn",
]

[tool.setuptools.packages.find]
include = ["src*"]
